Database Rebuild | unauthorized component init | GetUserEvents init | RoleGuard

+++ viko-api/viko-api/Migrations/VikoDbContextModelSnapshot.cs
@@ -75,8 +75,8 @@ namespace viko_api.Migrations
                         .HasColumnType("bigint")
                         .HasColumnName("Entity_Id");
 
-                    b.Property<long>("EventStatusId")
-                        .HasColumnType("bigint")
+                    b.Property<int>("EventStatusId")
+                        .HasColumnType("int")
                         .HasColumnName("Event_Status_Id");
 
                     b.Property<DateTime>("FinishDate")
@@ -142,11 +142,11 @@ namespace viko_api.Migrations
 
             modelBuilder.Entity("viko_api.Models.Entities.EventStatus", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Status")
                         .IsRequired()
@@ -162,17 +162,17 @@ namespace viko_api.Migrations
                     b.HasData(
                         new
                         {
-                            Id = 1L,
+                            Id = 1,
                             Status = "Open"
                         },
                         new
                         {
-                            Id = 2L,
+                            Id = 2,
                             Status = "Closed"
                         },
                         new
                         {
-                            Id = 3L,
+                            Id = 3,
                             Status = "Finished"
                         });
                 });
